warning in trap.c that recently appeared
src/trap.c(5410): warning C4389: '==': signed/unsigned mismatch

+++ src/trap.c
@@ -5407,7 +5407,7 @@ count_traps(int ttyp)
     struct trap *trap = g.ftrap;
 
     while (trap) {
-        if (trap->ttyp == ttyp)
+        if ((int) trap->ttyp == ttyp)
             ret++;
         trap = trap->ntrap;
     }
